requirements.txt
flask
python-dotenv
supabase
google-generativeai
